task,respawn: fixing separated init FSM vs respawn mode
