+++ kernel/src/syscalls/sysgate_exit.c
@@ -12,6 +12,7 @@ stack_frame_t *gate_exit(const stack_frame_t *frame, uint32_t result)
     taskh_t current = sched_get_current();
     taskh_t next;
     stack_frame_t *next_frame = NULL;
+    next = sched_elect();
 
     if (unlikely(mgr_task_set_jobreturn(current, result) != K_STATUS_OKAY)) {
         /* should never fail! */
@@ -35,25 +36,34 @@ stack_frame_t *gate_exit(const stack_frame_t *frame, uint32_t result)
             /* if this job is flagged to panic on exit, panic as requested */
             panic(PANIC_HARDWARE_UNEXPECTED_MODIFICATION);
         }
+        /* now electing a new job, sched_elect() never fails */
         /*
-         * requiring restart job are reswpaned (with a newly created handle)
-         *  Note that this job is scheduled, but not automatically elected
+         * As sched_elect() is called, current job has been removed from the scheduler list.
+         * From now, if the job requires restarting, it can be rescheduled again.
+         *
+         * Note that this job is scheduled, but not automatically elected
          */
         if (meta->flags.exit_mode == JOB_FLAG_EXIT_RESTART) {
+            /*
+             * respawn failure can only be triggered by a critical security failure emitted by
+             * the security manger. This behavior is treated as a critical failure, and is handled
+             * by a panic() at upper layer.
+             *
+             * note that respawning a job reschedule it with a new handler, but do not elect it immediately
+             */
             if (unlikely(task_respawn_job(current) != K_STATUS_OKAY)) {
                 panic(PANIC_KERNEL_INVALID_MANAGER_RESPONSE);
             }
+        } else {
+            /* no syscall return code here, as the job is **never** reelected.
+             * in order to ensure that such reelection do not happen, the task syscall return
+             * code value is set to NON_SENSE, generating a voluntary panic() if elected again
+             * For just respawned jobs, overall user data has been remapped, so this sysreturn value
+             * will be automatically cleared during the first system call made by the newly respawned job
+             */
+            mgr_task_set_sysreturn(current, STATUS_NON_SENSE);
         }
     }
-    /* no syscall return code here, as the job is **never** reelected.
-     * in order to ensure that such reelection do not happen, the task syscall return
-     * code value is set to NON_SENSE, generating a voluntary panic() if elected again
-     * For just respawned jobs, overall user data has been remapped, so this sysreturn value
-     * will be automatically cleared during the first system call made by the newly respawned job
-     */
-    mgr_task_set_sysreturn(current, STATUS_NON_SENSE);
-    /* now electing a new job, sched_elect() never fails */
-    next = sched_elect();
     if (unlikely(mgr_task_get_sp(next, &next_frame) != K_STATUS_OKAY)) {
         panic(PANIC_KERNEL_INVALID_MANAGER_RESPONSE);
     }
